Collapse GradientSamplerBase into NUTSSampler
GradientSamplerBase was originally introduced to share warmup adaptation
between HMCSampler and NUTSSampler. With pure HMC removed, NUTS is the
only descendant. The intermediate base class is dead weight: it adds an
inheritance layer with no payoff and uses misleading naming (the "shared
base for gradient-based MCMC" now serves a single sampler).

This commit folds the gradient scaffolding directly into NUTSSampler:

- step() now contains the Stage-3c restart, mass-matrix update orchestration,
  step-size heuristic re-run, and the leapfrog/RATTLE call as one body.
- initialize() / do_initialize() and the NUTSAdaptationController member
  (renamed nuts_adapt_) move into NUTSSampler as private members.
- The protected do_gradient_step() virtual hook goes away; NUTSSampler
  dispatches to do_unconstrained_step() / do_constrained_step() inline.

Hierarchy is now flat:
  SamplerBase  (abstract: step, initialize, has_nuts_diagnostics)
  ├── MetropolisSampler   (forwards to model.do_one_metropolis_step)
  └── NUTSSampler         (gradient scaffolding + leapfrog/RATTLE)

If a second gradient sampler returns later, re-introducing the base class
is mechanical. Per the design notes in
dev/plans/backlog/post-hmc-sampler-cleanup.md, no symmetric do_step() hook
is added: MetropolisSampler has nothing to scaffold, and the NUTS
adaptation controller's per-sampler ownership is principled (not symmetric
with the model-side MetropolisAdaptationControllers).

src/models/base_model.h doc comments updated: "Used by GradientSamplerBase"
→ "Used by NUTSSampler".

No behavior change; restructure only. All 6272 tests pass.
